requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
Werkzeug==2.3.7
SQLAlchemy==2.0.21
PyJWT==2.8.0
python-dotenv==1.0.0
cryptography==41.0.4
gunicorn==21.2.0
pymysql==1.1.0
Markdown==3.5
Pillow==10.0.1
email-validator==2.0.0.post2
WTForms==3.0.1